viewmodel for addVehicle

+++ src/Client/Model/Vehicle.java
@@ -10,11 +10,11 @@ public class Vehicle
   private int year;
   private String typeOfGearbox;
   private String typeOfFuel;
-  private int price;
+  private double price;
   private String status;
 
   public Vehicle(String licensePlate,int enginesPower,String type,String make,String model,int year,
-                 String typeOfGearbox, String typeOfFuel, int price){
+                 String typeOfGearbox, String typeOfFuel, double price){
 
     this.licensePlate = licensePlate;
     this.enginesPower = enginesPower;
@@ -105,7 +105,7 @@ public class Vehicle
     this.price = price;
   }
 
-  public int getPrice()
+  public double getPrice()
   {
     return price;
   }
